Add folder for awsS3 storage and add upload file test
